implement spring Jpa, enity and repository classes created

+++ backend/src/main/java/com/codecool/men/controller/dto/NewNoteDTO.java
@@ -3,6 +3,6 @@ package com.codecool.men.controller.dto;
 public record NewNoteDTO(
         String title,
         String text,
-        int userId
+        long userId
 ) {
 }
